requirements.txt
ragas
pandas
datasets
openpyxl
langchain-gigachat
langchain-openai
sentence-transformers
faiss-cpu
